requirements.txt
gevent==0.13.8
wsgiref==0.1.2
python-memcached==1.48
simplejson==3.1.2